+++ src/Schema/Blueprint.php
@@ -125,9 +125,10 @@ class Blueprint extends BaseBlueprint
     /**
      * @param  string  $column
      * @param int|null $length
+     * @param bool $fixed
      * @return ColumnDefinition
      */
-    public function binary($column, $length = null)
+    public function binary($column, $length = null, $fixed = false)
     {
         $length = $length ?: Builder::$defaultBinaryLength;
 
@@ -138,10 +139,9 @@ class Blueprint extends BaseBlueprint
      * @param string $column
      * @param int $total
      * @param int $places
-     * @param bool $unsigned
      * @return ColumnDefinition
      */
-    public function decimal($column, $total = 38, $places = 9, $unsigned = false)
+    public function decimal($column, $total = 38, $places = 9)
     {
         if ($total !== 38) {
             $this->markAsNotSupported('decimal with precision other than 38');
@@ -151,11 +151,7 @@ class Blueprint extends BaseBlueprint
             $this->markAsNotSupported('decimal with scale other than 9');
         }
 
-        if ($unsigned) {
-            $this->markAsNotSupported('unsigned decimal');
-        }
-
-        return parent::decimal($column, $total, $places, $unsigned);
+        return parent::decimal($column, $total, $places);
     }
 
     /**